Tutorial 1 - spike encoding
did the first tutorial for snntorch for spike encoding. Imo it seems like the actual 'useful' and practical choices is rate encodinga and maybe mix in event encoding. Idea being that rate encoding hard carries still images and whatnot, while event encoding is for differences and stuff to notice them better. Because when you stay unmoving it's not like you forget what you see, no, you remember it indefinitely. But at the same time you spot changes in vision really easily as well. So I guess some kind of change in a mostly still frame should be easy to guess, but a moving frame would largely be carried by rate encoding? idk. I'll stick to rate encoding for now, seems much more realistic. As for latency encoding I mean sure it is technically more grounded in truth but it's literally only useful for images, as for videos the much more realistic simulation is rate encoding